projects/lapaz-crawl-20260305170615/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
supabase>=2.10.0
openai>=1.50.0
python-dotenv>=1.0.1
selenium>=4.25.0
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
